requirements.txt
# Backend
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Document parsing
pdfplumber==0.10.3
python-docx==0.8.11

# NLP & ML
spacy==3.7.2
sentence-transformers==2.2.2
keybert==0.8.1
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
numpy==1.24.3
scikit-learn==1.3.2
scipy==1.11.4
torch==2.1.2

# Frontend
streamlit==1.28.1
plotly==5.17.0
requests==2.31.0

# Utilities
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
